add kpi

+++ ASP.NET - MVC/Farsi/10.0.0.3/AspnetCoreMvcStarter/Models/CrowdFunding/ShipmentStatistics.cs
@@ -2,7 +2,7 @@ namespace AspnetCoreMvcStarter.Models.CrowdFunding
 {
   public class ShipmentStatistics
   {
-    public int[] RequiredCapital { get; set; }
+    public long[] RequiredCapital { get; set; }
     public int[] InvestedCapital { get; set; }
   }
 }
